requirements.txt
PyQt5==5.15.9
pytesseract==0.3.10
keyboard==0.13.5
Pillow==10.2.0
pyttsx3==2.90
pywin32==306
numpy==1.24.3
d3dshot==0.16.0 